Set automation interval to something much higher
> `imageupdateautomation/apps-jenkins.flux-system`
> no updates made

This is quite noisy once every minute. It will send the same alerts when
it is triggered via news from the ImageRepository resource, so turn up
the interval from 1m0s to something sensible, so if there are no updates
for a week we still get the notification, only once a week.