Poll jenkins:lts image from docker.io/kingdonb/
Now that the webhook for ImageRepo is in place it is no longer necessary
to poll the Docker hub index with any frequency. We can increase this
reconcile interval to a very high number, just to ensure that it happens
with some periodicity even if the webhook receiver is not functioning.