separate internal and public ingresses
Services are internal unless they select the 'public' ingress class
Make all services internal by default (unless they are webhooks)

Keycloak is also internal

Update all ingress classes in helmreleases: gitea (disabled), jenkins,
the gitno.hephy.pro webhook ingress (public), update all clusterissuers
to be public (aha! this is cool... internal services with access to the
clusterissuer can request a certificate, and for cert-manager, it is
public but the application itself remains internal only! So cool...)

Add a new internal-ingress and designate the old ingress as public. The
internal should come up first, and get the ip address 10.17.12.201, the
public ingress comes up second and is at 10.17.12.202 the public DMZ.

See:

$ host 10.17.12.201
201.12.17.10.in-addr.arpa domain name pointer nginx-internal.nerdland.local.
$ host 10.17.12.202
202.12.17.10.in-addr.arpa domain name pointer nginx-dmz.nerdland.local.

(Another public-facing HTTP server is the front-end, which is not
controlled from inside this cluster, so new ingresses here that claim to
be public are in fact **not** automatically public facing.)

The internal-until-approved type of administrative security posture. 👍

It would not make sense to have Jenkins be automatically public facing,
and there are some not-so-graceful failure modes where if the config is
not loaded successfully, anonymous users are authorized, and just no!

You do not want one of these in your internal network that is connected
to the public internet. Capital letters N-O No. Absolutely not. No way.
It could be a ticking time bomb unless and until it is verified as
properly configured, or it is kept as an internal service.