Jenkins itself provides a mechanism for separating
webhook which is meant to be a public-facing concern, from Jenkins the
internal service which is meant to be a high-security environment that
stores secrets and requires authorized users only for high-stakes tasks.

Jenkins configures its own webhooks dynamically through the Github Org
configuration that we have done, so this is a really helpful addition!